+++ src/ambition/version_details.cc
@@ -12,6 +12,7 @@ std::vector<VersionUpdates> versionDetails = {
       {
         .heading = _("Added"),
         .items = {
+          _("Added clipboard keys (Ctrl+C, Ctrl+X, Ctrl+V) to input fields, such as the address of the \"Enter Address\" multiplayer option, map ID, and name fields."),
           _("Automatically copy the map ID to the clipboard when opening the in-game menu, allowing for easy copy-pasting of the map ID."),
           _("Automatically paste into the join address field, allowing for easy joining of games when using the \"Enter Address\" option."),
           _("Added a What's New screen, shown on the first startup of each new version."),