+++ CHANGELOG.md
@@ -12,6 +12,8 @@ speaking, considered to be the player's experience.
 
 ### Added
 
+- Added clipboard keys (Ctrl+C, Ctrl+X, Ctrl+V) to input fields, such as the
+  address of the "Enter Address" multiplayer option, map ID, and name fields.
 - Automatically copy the map ID to the clipboard when opening the in-game menu,
   allowing for easy copy-pasting of the map ID.
 - Automatically paste into the join address field, allowing for easy joining of